requirements.txt
django
djangorestframework
pre-commit
django-cors-headers